saving eprom

+++ convertToHours.ino
@@ -3,7 +3,7 @@
   unsigned long minutes = seconds /60; 
   unsigned long hours = minutes/60;
 
-  return hours;
+  return (int) hours;
 }
